Continue working on availability, address Vin's comment on #89
